debianrepochecker: use libapt
The Google Chrome repo currently has an
out-of-date copy of Packages.bz2. Compare and contrast:

    $ wget --quiet -O- https://dl.google.com/linux/chrome/deb/dists/stable/main/binary-amd64/Packages.bz2 | bunzip2 | grep -A1 'Package: google-chrome-stable'
    Package: google-chrome-stable
    Version: 70.0.3538.77-1
    $ wget --quiet -O- https://dl.google.com/linux/chrome/deb/dists/stable/main/binary-amd64/Packages | grep -A1 'Package: google-chrome-stable'
    Package: google-chrome-stable
    Version: 71.0.3578.98-1

This sounds like a repository bug, but actually it's a client bug. The
Release file for the stable distribution only lists Packages.gz and
Packages, not Packages.bz2:

    $ wget --quiet -O- https://dl.google.com/linux/chrome/deb/dists/stable/Release
    Origin: Google LLC
    Label: Google
    Suite: stable
    Codename: stable
    Version: 1.0
    Date: Wed, 09 Jan 2019 18:36:46 UTC
    Architectures: amd64
    Components: main
    Description: Google chrome-linux software repository
    MD5Sum:
     1d8607371ba7b15efff49a271c043797 4457 main/binary-amd64/Packages
     73d44b63e4c7d55328a6b18b852f1161 1113 main/binary-amd64/Packages.gz
     156e5ea7a0c6bed5973a68a45e546dc9 151 main/binary-amd64/Release
    SHA1:
     0523ad118b5d8494a3267fa7a3913e943918a99e 4457 main/binary-amd64/Packages
     1551e8b260f248904d2ec8d2dec4826a3d8cd330 1113 main/binary-amd64/Packages.gz
     0f4348c2d4d7cc1f8e59b5934d87f1ca872f6e34 151 main/binary-amd64/Release
    SHA256:
     4dcbe011302d736415c90c03a5c9ba234563aec96e2eac7c8231fe20f1034c1c 4457 main/binary-amd64/Packages
     82d8fd921f6129bf5ff02be3fb883206506941691d2fd0c3805c13fc883354c2 1113 main/binary-amd64/Packages.gz
     c1e3c9318381862306adcdc4fd4fe2d85be8aa4c4f3dcbb40fce80413f588286 151 main/binary-amd64/Release

Apt itself fetches the Release file to work out which package index URL
to use. You can check with `sudo apt update -o Debug::Acquire::http=1`:
it fetches the Release file above, and then fetches Packages.gz.

The previous version of this code checked Packages.xz, Packages.bz2, and
Packages, in that order, without first checking the Release file for the
distribution. Instead of open-coding this, use libapt (via the apt_pkg
module) to get exactly the same code that apt itself uses. The code in
_load_repo() is derived from an Endless-internal project, originally
written by Dan Nicholson.

There's another behaviour change here: the previous code would report no
error in the case where the package is not found in the repo, whereas
this version will throw an exception at:

    package = cache[package_name]

I think this is preferable.

https://phabricator.endlessm.com/T25134