urlchecker: check remote checksum and size
https://github.com/flathub/com.adobe.Flash-Player-Projector has:

  {
    "type" : "extra-data",
    "filename" : "flash_player_sa_linux.x86_64.tar.gz",
    "url" : "https://fpdownload.macromedia.com/pub/flashplayer/updaters/32/flash_player_sa_linux.x86_64.tar.gz",
    "sha256" : "997b857438a52a3afa3c8072bca60a2ff1cacd069b17662197d17e554a18e384",
    "size" : 8503904
  },

which is annoyingly half-way between being a rotating URL (the content
changes as minor versions are released) and not (the URL changes on
major version updates). Previously this checker would assume that it is
valid, even if the content has changed.

To avoid downloading the same data twice in the case where a
rotating-url source is up-to-date, make RotatingURLChecker set the state
to VALID or BROKEN if it's sure one way or the other, and make
ManifestChecker stop once ExternalData.state becomes something other
than UNKNOWN.