Dockerfile: clean up /var/cache/apt/archives/partial
This directory is empty and looks like this:

    drwx------ 2 _apt root 4096 Jun 17 22:00 /var/cache/apt/archives/partial

Since on Jenkins the container runs as a non-root user, this directory
is opaque. apt_pkg for some reason tries to clean it up (even though
it's been told to use a different directory) and prints some warnings:

    rm: cannot remove '/var/cache/apt/archives/partial/*.deb': Permission denied
    rm: cannot remove '/var/cache/apt/archives/partial/*.deb': Permission denied
    rm: cannot remove '/var/cache/apt/archives/partial/*.deb': Permission denied

Clean it up at image build time.

https://phabricator.endlessm.com/T19948