debianrepochecker: wire up progress to logging
If you run the tests with '-b' to buffer output:

    python3 -m unittest -bv tests/test_*

You would previously see lots of junk in the terminal:

    Error in function start
    Error in function pulse
    Error in function pulse
    [...]

Followed by, if the test fails (at which point unittest shows the
buffered output):

    Traceback (most recent call last):
      File "/usr/lib/python3/dist-packages/apt/progress/text.py", line 136, in start
        self._winch()
      File "/usr/lib/python3/dist-packages/apt/progress/text.py", line 142, in _winch
        if hasattr(self._file, "fileno") and os.isatty(self._file.fileno()):
    io.UnsupportedOperation: fileno

I think this is a bug in python-apt. fileno() is documented to exist but
raise an exception if the IO object does not use a file descriptor.
https://docs.python.org/3/library/io.html?highlight=iobase#io.IOBase.fileno

However, it turns out to be pretty easy in practice to wire this up to
a logger, with only a little bit of ill-gotten knowledge about
text.AcquireProgress's internals.