urlchecker: don't follow redirects for non-rotating URLs
b249ee4 introduced a regression: if the URL is stable, but happens to
be a redirect (to a CDN, for example), the checker would update it to
the CDN URL. See
https://github.com/flathub/com.visualstudio.code/pull/99 for an example
of this happening in practice.

Fix this by restoring the old behaviour of only using the post-redirect
URL if is_rotating is true. I think the logic is a little hard to follow
now, but I'm not sure how to write it more cleanly.

Add tests for both the rotating and the non-rotating cases.