Use main manifest filename when building appdata filename
This fixes an issue for Firefox which uses separate manifests
for ext-data sources and thus was not getting the appdata updated
as the checker would try to use the name of the manifest that was
changed (the separate one) to build the appdata filename and thus
fail (skip the appdata update) as the appdata was not found.

Lets use the name of the main manifest passed to the checker instead.

Signed-off-by: Andre Moreira Magalhaes <andre@endlessm.com>