firefoxchecker: HEAD, don't GET, each file
We get the list of files from a file called SHA256SUMS, which
unsurprisingly contains the SHA256 of each file. So we don't need to
fetch them: we just need to HEAD them to determine their Content-Length
and timestamp.

On my machine & network, test_firefoxchecker used to take 2m28s; it now
takes 16s.