utils: remove unneeded variable assignments
All these are assigned later in the function.

In a language like Python where variables don't need to be declared at
the top of the function, I think it is better not to initialize them
with bogus values: this way, the interpreter will raise an exception if
you follow a code path where it is used before being assigned.