checker: don't crash when sources are added or removed
Added or removed external data do not have data.new_version set –
defensible since the datum is itself new. This caused the checker to
crash when, in the new release, three new languages were added to
Firefox.

In any case, this code path should not have been reached because we
attemp to generate a message of the form "Added _" or "Removed _". Fix
that.