appdata: switch to defusedxml
The [xml.sax][] documentation begins with a dire warning:

> The xml.sax module is not secure against maliciously constructed data.
> If you need to parse untrusted or unauthenticated data see [XML
> vulnerabilities][].

The linked page shows that xml.sax is vulnerable to, at least, two
entity expansion attacks, and recommends defusedxml as a drop-in
replacement.

While the XML handled by this library is not exactly *untrusted*, it is
not exactly *trusted* either. We don't expect appdata files to be in the
business of defining new XML entities, so drop in defusedxml to prevent
this.

Thanks to Seppo Yli-Olli for pointing out this issue.

[xml.sax]: https://docs.python.org/3/library/xml.sax.html
[XML vulnerabilities]: https://docs.python.org/3/library/xml.html#xml-vulnerabilities