urlchecker: Strip query string from redirect target
This script is spamming the Zoom repo with patches like this:

-"url": "https://d11yldzmag5yn.cloudfront.net/prod/3.5.372466.0322/zoom_x86_64.tar.xz",
+"url": "https://d11yldzmag5yn.cloudfront.net/prod/3.5.372466.0322/zoom_x86_64.tar.xz?_x_zm_rtaid=muDd1uOqSZ-xUScZF698QQ.1585134521724.21e5ab14908b2121f5ed53882df91cb9&_x_zm_rhtaid=732",

This appears to be a marketing tag, and as such changes every time the
bot scrapes the homepage.

To avoid this, strip all query strings when following a rotating-url.

In principle, there might be situations where the query string has to be
preserved, so this would need to be configurable. In practice, I expect
that only the test case changed in this patch will be affected.

Fixes #78