Re-add support for running container with arbitrary internal uid:gid
We pass the user and group ids when building the image used by
CI/Jenkins as a workaround to make sure the user running the container
exists inside of it (see Jenkinsfile/Dockerfile). This is needed because
jenkins runs the container with the same user as the host
(with '-u <uid>:<gid>') but 'git' fails with 'No user exists for uid ...'
if the user doesn't exist inside the container.

See also https://issues.jenkins-ci.org/browse/JENKINS-47026.