Exit with success if PR can be opened
Previously, this script would exit(0) if some sources were out of date
but we were able to open a PR to update them. We want this behaviour
because the owner of whatever CI job runs the checker should not get a
failure email if the script opens a PR: that's a success for the
script! The script should only indicate failure if it is not able to
carry out the task it was asked to perform (edit the source tree, open a
PR, etc.).

f4cddf5d12579c39e1fac58f1730d5f553a4c58b changed this behaviour. Change
it back.