chromiumchecker: Add checker for Chromium components
This is needed to be able to auto-update the versions for the LLVM
toolchain components in sync with Chromium itself. It also can update
the Chromium tarball; the Anitya checker could already technically do
this, but this code needs to parse the versions anyway, so actually
updating the tarball in sync is a minor addition.