chromiumchecker: Don't attempt to resolve commits
At the time of writing, I didn't realize fetch_remote() actually is
intended to find the commit value of a tag or branch; if neither is
given, it assumes you want the master branch. In other words, this would
end up discarding the commit revision that it was given and just grab
LLVM head.