Add --[always|never]-fork CLI parameters
I am trying to use this in GitHub Actions, with the default GITHUB_TOKEN
provided to the action. However, origin_repo.permissions.push is False.
Some digging suggests that this may be an artefact of the kind of token
being used here.[0] However, if we just go ahead and push and create the
CR directly to the repo, it works fine.

So, add a --never-fork CLI parameter to override the repository
permissions check, and an --always-fork parameter for symmetry. The
default is still to check the repository permissions. (Of course if you
pass --never-fork and you really don't have permission to push or create
PRs, you'll hit an error a few lines later.)

[0] https://github.com/endlessm/eos-google-chrome-app/pull/96