main: Don't check permissions before automerge
As described in the previous commit, when using the GITHUB_TOKEN
provided by GitHub Actions, the permissions check incorrectly claims
that we do not have push access to the repository.

Instead of checking the repository permissions, in the case where we
would like to automerge, just try to do it. The worst that happens is
that the pr.merge() call fails.

I think it is preferable not to tie this to the --never-fork
command-line argument because you might conceivably want to
--always-fork but still automatically merge. We already have a mechanism
to override the automerge behaviour in flathub.json.