lib/utils: Catch type errors on loose version comparison
Some software[0] does not strictly use numbers or semantic versioning
for labeling tags/releases in git. Trying to filter out these versions
would lead to TypeError due to comparison between int and str (str being
the non-compliant version string). Catch this error and due to no
reliable of checking the version, mark it as a non-match.

[0] https://release-monitoring.org/project/13289/