Add support for allowing selective manifest updates If `--require-important-update` is passed to CLI, and a `is-important: true` and/or `is-main-source: true` source is present in external data, only update the manifest if such "important" external data received an update. Manifests not meeting both conditions will not be affected by this change. f-e-d-c would continue making updates as before for all modules.
Use `--require-important-update` CLI argument to gate this feature. This ensures e.g. copy-pasted modules alone cannot accidentally activate this feature.

Require "important" sources to be set explictly, since there is no clear way to automatically detect which sources are "important".
Even if there was a way of detecting "important" sources, the possibility that f-e-d-c detects wrongly could be confusing to debug.

Don't reuse the appdata updater's guess that the last source is the "main" source, (done by appdata in case `is-main-source: true` is not found).
It could be confusing if the feature was enabled and f-e-d-c arbitrarily started only updating the manifest if the last source was updated.

In the future, this change could be extended with an "always-update-for-cves" setting in addition to the require-important-update setting.
Such a setting could update the manifest if a source with CVEs received an update, regardless if the updated sources are set as is-important/is-main-source.