Consistently default stable-only to true
Previously, gnomechecker would default stable-only to true, but
anityachecker and pypichecker defaulted it to false.

true is the sensible default, in my opinion: it is more common to want
stable release only than it is to track prereleases.

Update the test suite to exercise all three paths: true, false and
unset. (We don't actually make any assertions about what the default
is; it is hard to do so without a mock anitya.)