Detect content type from magic number of first chunk
To work around an issue where SourceForge sometimes returns an HTML
error page with a successful response code, we started checking the
content-type of URLs we fetch and rejecting certain text formats in some
contexts.

Sadly some servers serve source-code archives with an incorrect
Content-Type header.

To work around this, ignore the Content-Type header and use the 'magic'
library to sniff the content type of the returned data.

Fixes #245