tests: Stop using ghostscript in tests
One of the tests exercises the ability to use parsed components of
versions in URL patterns.

Ghostscript uses version numbers of the form 10.01.0. Since the
components are treated as digits, the $version1 variable is set to "1"
not "01", and so this test fails.

See https://github.com/flathub/flatpak-external-data-checker/issues/360.

To work around this bug to unbreak the test suite, switch to a different
module.