Dockerfile: Check requirements.txt installs nothing
We want all dependencies to have been handled by the packages in
dependencies.apt.txt.

pip doesn't have a built-in way to say "check this is a no-op" but we
can do it by asking it to report on a dry-run and checking that the
installation report has an empty list of packages to be installed.

https://pip.pypa.io/en/stable/reference/installation-report/