htmlchecker: allow specifying error handling on encoding error
I need to fetch a binary encoded file, which contains the update name. The file is binary and just contains some readable strings. The problem is, that it already fails with `Error querying for new versions: 'utf-8' codec can't decode bytes in position` thus I am not able to apply any regexes on it. Now I can set an attribute called `encoding-error` to `ignore`. This is kinda hacky.