gitchecker: Use Checker._get_pattern helper
Rather than asserting when given the wrong number of match
groups, this helper raises a descriptive exception. This is both more
helpful to the developer who has incorrectly specified the tag-pattern,
but also seems to prevent an issue where an assertion failure causes the
asyncio mainloop to hang.

Fixes #418

Thanks to Tobias Micheler <tobiasmicheler@outlook.de> for finding the
issue!