main: Be more resistant to git commit failures
If the default git commit fails try a sanitised option that does not
read any config files, skips GPG signatures/hooks and uses a on-the-fly
calculated email.

Only the email needs to be calculated and manually set because on
systems with no `/etc/mailname` or `/etc/hostname` such as Flatpak
runtime it generates one ending with `none` that it considers bogus.
Committing using that will fail again.

[1]: https://github.com/git/git/blob/c3ebe91b40c23a1b70dba36383a23016711d8bc0/ident.c#L505-L513
[2]: https://github.com/git/git/blob/c3ebe91b40c23a1b70dba36383a23016711d8bc0/ident.c#L157-L159
[3]: https://github.com/git/git/blob/c3ebe91b40c23a1b70dba36383a23016711d8bc0/ident.c#L135-L142