main: Set module names in the PR title
Choose the "best" title with most module names while making sure it does
not cross 70 characters. The limit is to prevent too much verbosity in
commit messages and PR titles when a large set of modules are getting
updated.

If it crosses 70 characters, start deducting module names and add the
number of remaining updated modules.